refactor: 점수 자료형 double로 변경

+++ db/db/mysql/init/init.sql
@@ -2,7 +2,7 @@ use Janggi;
 
 CREATE TABLE player (
     id  INT NOT NULL AUTO_INCREMENT,
-    score INT NOT NULL,
+    score DOUBLE NOT NULL,
     team ENUM ('HAN', 'CHO') NOT NULL,
     PRIMARY KEY (id)
 );
